Move runLoader to own module
- Define LoaderRunner type
- Implement constructor to take a TruffleDB instance
- Remove implementation from TruffleDB instance
